feat: add deepsearch debug_report
